// src/app/system/hr/models/interfaces/annual.ts
export interface AnnualViewModel {
  id: string;
  name: string;
  maxRequestNum: number;
  confirmationLayerNum: number;
}
export class searchRequestTypeViewModel {
  Name: string;
}
export class deleteRequestTypeViewModel {
  id: string;
}
export interface createRequestTypeViewModel {
  id: string;
  name: string;
  maxRequestNum: number;
  confirmationLayerNum: number;
}